Add option to specify rebuild image via a file
You can add a file on the node to trigger the rebuild.
We look for a file with an image uuid in the following location:
/var/spool/slurm/REBUILD_IMAGE_UUID

This is required when you can't specify a reboot reason,
such as when using a job to trigger the rebuild.

Note the file takes precidence over an sinfo reason,
in part because it means we don't need to reply on
sinfo being installed in a specific location in this
case.